[🐛fix]: storybook에서 ThemeProvider 적용 안되는 오류 수정
- storybook에서 ThemeProvider의 theme이 적용이 안되는 오류 발견
- .storybook/preview.ts파일을 수정함으로 해결
